// Ai-Interview-Frontend/src/app/shared/interfaces/careers.interfaces.ts
export interface CareerHero {
  badge: string;
  title: string;
  highlightedText: string;
  description: string;
  image: string;
}

export interface WorkplaceImage {
  image: string;
  title?: string;
  subtitle?: string;
}

export interface Benefit {
  icon: string;
  title: string;
  description: string;
}